add logmanager
