+++ tests-dev/logmanager.py
@@ -182,10 +182,16 @@ Result: {result}
         for app, jobs in rt_yaml.items():
             for key, val in jobs.items():
                 if key == "build":
-                    self.parse_compile_log(app, val)
+                     rt_compiler = val['compiler']
+                     self.parse_compile_log(app, val)
                 elif key == "tests" and not self.compile_only:
                     for test in val:
                         case, config = get_testcase(test)
-                        self.parse_test_log(case, config, val["compiler"])
+                        self.parse_test_log(case, config, rt_compiler)
 
         self.finalize()
+      
+def run_regression_logging():
+    from logmanager import RegressionLogManager  # adjust import if needed
+    logger = RegressionLogManager()
+    logger.run()
