+++ tests-dev/ufs_test_utils.py
@@ -7,157 +7,6 @@ import shutil
 import subprocess
 from datetime import datetime
 
-def get_timestamps(path):
-    """Obtain experiment starting and ending time marks through file timestamps
-
-    Args:
-        path (str): experiment log directory
-    Returns:
-        str: experiment starting and ending time strings
-    """
-    dir_list = os.listdir(path)
-    dt = []
-    for f in dir_list:
-        m_time = os.path.getmtime(path+f)
-        dt.append(datetime.fromtimestamp(m_time))
-    dtsort=sorted(dt)
-    return str(dtsort[0]),str(dtsort[-1])
-
-def finalize_regression_log(env, run_logs, compile_pass, compile_nr, job_nr, pass_nr, fail_nr, failed_list, test_changes_list):
-    filename = os.path.join(env.pathrt, f"logs/RegressionTests_{env.machine_id}.log")
-    write_logfile(filename, "a", output=run_logs)
-
-    test_start_time, test_end_time = get_timestamps(f"./logs/log_{env.machine_id}/")
-    start_time = datetime.strptime(test_start_time.split('.')[0], "%Y-%m-%d %H:%M:%S")
-    end_time = datetime.strptime(test_end_time.split('.')[0], "%Y-%m-%d %H:%M:%S")
-    elapsed = end_time - start_time
-    hours, remainder = divmod(elapsed.total_seconds(), 3600)
-    minutes, seconds = divmod(remainder, 60)
-
-    elapsed_time = f"{int(hours):02}h:{int(minutes):02}m:{int(seconds):02}s"
-    synop_log = f"""
-SYNOPSIS:
-Starting Date/Time: {test_start_time}
-Ending Date/Time: {test_end_time}
-Total Time: {elapsed_time}
-Compiles Completed: {int(compile_pass)}/{int(compile_nr)}
-Tests Completed: {int(pass_nr)}/{int(job_nr)}
-"""
-    write_logfile(filename, "a", output=synop_log)
-
-    if int(fail_nr) == 0:
-        if os.path.isfile(test_changes_list):
-            delete_files(test_changes_list)
-        open(test_changes_list, 'a').close()
-        comment_log = f"""
-NOTES:
-A file test_changes.list was generated but is empty.
-If you are using this log as a pull request verification, please commit test_changes.list.
-Result: SUCCESS
-====END OF {env.machine_id} REGRESSION TESTING LOG====
-"""
-    else:
-        with open(test_changes_list, 'w') as listfile:
-            for line in failed_list:
-                listfile.write(f"{line}\n")
-        comment_log = f"""
-NOTES:
-A file test_changes.list was generated with list of all failed tests.
-You can use './rt.sh -c -b test_changes.list' to create baselines for the failed tests.
-If you are using this log as a pull request verification, please commit test_changes.list.
-Result: FAILURE
-====END OF {env.machine_id} REGRESSION TESTING LOG====
-"""
-    write_logfile(filename, "a", output=comment_log)
-
-def process_test_log(machine_id, test_name, compiler, config):
-    """
-    Process a single test job and return its log string and pass status.
-    """
-    test_id = f"{test_name}_{compiler}"
-    test_log_path = f"./logs/log_{machine_id}/rt_{test_id}.log"
-    timestamp_path = f"./logs/log_{machine_id}/run_{test_id}_timestamp.txt"
-    pass_check = f"Test {test_id} PASS"
-    maxs_check = "The maximum resident set size (KB)"
-    pass_flag = False
-    memsize = ""
-    time_log = ""
-
-    try:
-        with open(test_log_path) as f:
-            if pass_check in f.read():
-                pass_flag = True
-    except FileNotFoundError:
-        print(f"{test_log_path}: does not exist")
-        return f"FAIL -- TEST {test_id}\n", False, test_name
-
-    if pass_flag:
-        try:
-            with open(timestamp_path) as f:
-                timing_data = f.read().split('\n', 1)[0]
-                etime = int(timing_data.split(",")[4].strip()) - int(timing_data.split(",")[1].strip())
-                rtime = int(timing_data.split(",")[3].strip()) - int(timing_data.split(",")[2].strip())
-                etime_min, etime_sec = divmod(etime, 60)
-                rtime_min, rtime_sec = divmod(rtime, 60)
-                time_log = f" [{etime_min:02}:{etime_sec:02}, {rtime_min:02}:{rtime_sec:02}]"
-        except FileNotFoundError:
-            time_log = ""
-
-        try:
-            with open(test_log_path) as f:
-                for line in f:
-                    if maxs_check in line:
-                        memsize = line.split("=")[1].strip()
-                        break
-        except FileNotFoundError:
-            memsize = ""
-
-        return f"PASS -- TEST {test_id}{time_log} ({memsize} MB)\n", True, None
-    else:
-        return f"FAIL -- TEST {test_id}\n", False, test_name
-
-def process_compile_log(machine_id, compile_id, compiler, pathrt):
-    """
-    Process a single compile job and return its log string and pass status.
-    """
-    compile_log_path = f"./logs/log_{machine_id}/compile_{compile_id}.log"
-    timestamp_path = f"./logs/log_{machine_id}/compile_{compile_id}_timestamp.txt"
-    compile_err_path = f"{pathrt}/compile_{compile_id}/err"
-
-    compile_check1 = f"Compile {compile_id} Completed"
-    compile_check2 = "[100%] Linking Fortran executable"
-    compile_pass = False
-    warning_log = ""
-
-    try:
-        with open(compile_log_path) as f:
-            contents = f.read()
-            if compile_check1 in contents or compile_check2 in contents:
-                compile_pass = True
-            f.seek(0)
-            for line in f:
-                if 'export RUNDIR_ROOT=' in line:
-                    rundir_root = line.split("=")[1].strip()
-                    break
-            compile_err = f"{rundir_root}/compile_{compile_id}/err"
-        with open(compile_err) as ferr:
-            err_contents = ferr.read()
-            count_warning = err_contents.count(": warning #")
-            count_remarks = err_contents.count(": remark #")
-            warning_log = f" ({count_warning} warnings, {count_remarks} remarks)"
-        with open(timestamp_path) as flog:
-            timing_data = flog.read().split('\n', 1)[0]
-            etime = int(timing_data.split(",")[4].strip()) - int(timing_data.split(",")[1].strip())
-            btime = int(timing_data.split(",")[3].strip()) - int(timing_data.split(",")[2].strip())
-            etime_min, etime_sec = divmod(etime, 60)
-            btime_min, btime_sec = divmod(btime, 60)
-            time_log = f" [{etime_min:02}:{etime_sec:02}, {btime_min:02}:{btime_sec:02}]"
-        status = "PASS" if compile_pass else "FAIL"
-        return f"{status} -- COMPILE {compile_id}{time_log}{warning_log}\n", compile_pass
-    except FileNotFoundError:
-        print(f"{compile_log_path}: does not exist")
-        return f"FAIL -- COMPILE {compile_id}\n", False
-
 def update_testyaml(input_list):
     """Generates temporary test YAML based on list of tests received
